+++ lib/features/homepage/model/activities.dart
