--- lib/ui/screens/editprofile.dart
@@ -1,157 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:get/get.dart';
-
-import '../../color/colors.dart';
-import '../component/backarrow.dart';
-
-class Editprofile extends StatelessWidget {
-  const Editprofile({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Color(0xffF1F6FA),
-      appBar: AppBar(
-        backgroundColor: Colors.white,
-        elevation: 0,
-        leading: const Padding(
-          padding: EdgeInsets.all(10.0),
-          child: Backarrow(color: Color(0xffFF7CCB)),
-        ),
-        title: Text(
-          "Edit Profile".tr,
-          style: const TextStyle(
-              fontWeight: FontWeight.w700,
-              fontSize: 18,
-              color: Color(0xff000E3B)),
-        ),
-        actions: [
-          Padding(
-            padding: const EdgeInsets.all(10.0),
-            child: Icon(
-              Icons.settings,
-              color: AppColor.primary_blue,
-            ),
-          )
-        ],
-      ),
-      body: Container(
-        padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 20),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Container(
-              width: Get.width,
-              alignment: Alignment.center,
-              child: Stack(
-                alignment: AlignmentDirectional.center,
-                children: [
-                  Image.asset(
-                    "assets/images/perrybot/profile.png",
-                    width: 84,
-                    height: 84,
-                  ),
-                  Center(
-                    child: Icon(
-                      Icons.camera_alt,
-                      color: AppColor.yellow,
-                      size: 35,
-                    ),
-                  )
-                ],
-              ),
-            ),
-            const SizedBox(
-              height: 30,
-            ),
-            const Text(
-              "basic_info",
-              style: TextStyle(fontWeight: FontWeight.w700, fontSize: 14),
-            ),
-            const SizedBox(
-              height: 20,
-            ),
-            Container(
-              width: Get.width,
-              padding: const EdgeInsets.symmetric(horizontal: 15, vertical: 10),
-              decoration: const BoxDecoration(
-                  color: Color(0xffEBEBEB),
-                  borderRadius: BorderRadius.all(Radius.circular(5))),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(
-                    "full_name".tr,
-                    style: const TextStyle(
-                        fontSize: 10, fontWeight: FontWeight.w600),
-                  ),
-                  const SizedBox(
-                    height: 10,
-                  ),
-                  const Text(
-                    "June Allen",
-                    style: TextStyle(fontSize: 13, fontWeight: FontWeight.w500),
-                  ),
-                ],
-              ),
-            ),
-            const SizedBox(
-              height: 20,
-            ),
-            Container(
-              width: Get.width,
-              padding: const EdgeInsets.symmetric(horizontal: 15, vertical: 10),
-              decoration: const BoxDecoration(
-                  color: Color(0xffEBEBEB),
-                  borderRadius: BorderRadius.all(Radius.circular(5))),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(
-                    "username".tr,
-                    style: const TextStyle(
-                        fontSize: 10, fontWeight: FontWeight.w600),
-                  ),
-                  const SizedBox(
-                    height: 10,
-                  ),
-                  const Text(
-                    "@illojune",
-                    style: TextStyle(fontSize: 13, fontWeight: FontWeight.w500),
-                  ),
-                ],
-              ),
-            ),
-            const SizedBox(
-              height: 20,
-            ),
-            Container(
-              width: Get.width,
-              padding: const EdgeInsets.symmetric(horizontal: 15, vertical: 10),
-              decoration: const BoxDecoration(
-                  color: Color(0xffEBEBEB),
-                  borderRadius: BorderRadius.all(Radius.circular(5))),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(
-                    "email_address".tr,
-                    style: const TextStyle(
-                        fontSize: 10, fontWeight: FontWeight.w600),
-                  ),
-                  const SizedBox(
-                    height: 10,
-                  ),
-                  const Text(
-                    "Checkforjune@gmail.com",
-                    style: TextStyle(fontSize: 13, fontWeight: FontWeight.w500),
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
